# api/API/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api de blog du cours Api",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "sandbox-server"
        }
    ],
    "paths": {
...
        "/blog": {
            "get": {
                "summary": "List all blog messages",
                "description": "List all blog messages",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "limit permettant la pagination",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "apikey permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List all blog messages",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Blog[]",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Blog"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "apiKey missing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "apikey manquante",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "apikey manquante",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/blog/{id}": {
            "get": {
                "summary": "return a blog message",
                "description": "return a blog message",
                "operationId": "27e01b7270e2200f5bde1c9c52fe6225",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id du message de blog à récupérer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "apikey permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return a blog message",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Blog"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "apiKey missing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "apikey manquante",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "apikey manquante",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/blog/create": {
            "post": {
                "summary": "Create a blog message",
                "description": "Save a blog message in DB",
                "operationId": "6cf98894e728d7fc44dccd2bc636f3f0",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "apikey permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Message à enregistrer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pseudo": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Voici un message"
                                    },
                                    "titre": {
                                        "type": "string",
                                        "example": "Bonjour"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message de blog enregistré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/blog/{id}/update": {
            "put": {
                "summary": "Update blog",
                "description": "Update blog in Db",
                "operationId": "1cc25fc2547f0b7f8696cff771087213",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "apikey permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id du message de blog à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Nouvelles données du message de blog",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pseudo": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "Voici un message"
                                    },
                                    "titre": {
                                        "type": "string",
                                        "example": "Bonjour"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message enregistré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/blog/{id}/delete": {
            "delete": {
                "summary": "Delete blog",
                "description": "Delete blog in Db",
                "operationId": "3f28c394f40d3fae23c0f5e92d1dc306",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id du blog à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "apikey permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Suppression validée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Blog message supprimé"
                            }
                        }
                    }
                }
            }
        },
...
    }
}